fix(onvif): Stop clearing the device list
The current implementation assumes the list of devices needs to be
reset on every probe.  Unfortunately, this does not allow for probing
multiple network interfaces.  This change adds a method that allows
the caller to clear the list of devices (resetDeviceList()) and also
uses an object as the second parameter to startProbe.  This "options"
object has two fields: resetDeviceList which is used to direct this
method whether or not to clear the list of devices (default is true),
and address which is used to indicate the IP address/hostname to which
the probe message is to be sent (default is the multicast address).